+++ chain/client/src/client_actor.rs
@@ -364,131 +364,8 @@ impl Handler<NetworkClientMessages> for ClientActor {
                 NetworkClientResponses::NoResponse
             }
             NetworkClientMessages::StateResponse(state_response_info) => {
-                let shard_id = state_response_info.shard_id();
-                let hash = state_response_info.sync_hash();
-                let state_response = state_response_info.take_state_response();
-
-                trace!(target: "sync", "Received state response shard_id: {} sync_hash: {:?} part(id/size): {:?}",
-                    shard_id,
-                    hash,
-                    state_response.part().as_ref().map(|(part_id, data)| (part_id, data.len()))
-                );
-                // Get the download that matches the shard_id and hash
-                let download = {
-                    let mut download: Option<&mut ShardSyncDownload> = None;
-
-                    // ... It could be that the state was requested by the state sync
-                    if let SyncStatus::StateSync(sync_hash, shards_to_download) =
-                        &mut self.client.sync_status
-                    {
-                        if hash == *sync_hash {
-                            if let Some(part_id) = state_response.part_id() {
-                                self.state_sync_addr.do_send(
-                                    StateSyncActorRequests::ReceivedRequestedPart {
-                                        part_id,
-                                        shard_id,
-                                        hash,
-                                    },
-                                );
-                            }
-
-                            if let Some(shard_download) = shards_to_download.get_mut(&shard_id) {
-                                assert!(
-                                    download.is_none(),
-                                    "Internal downloads set has duplicates"
-                                );
-                                download = Some(shard_download);
-                            } else {
-                                // This may happen because of sending too many StateRequests to different peers.
-                                // For example, we received StateResponse after StateSync completion.
-                            }
-                        }
-                    }
-
-                    // ... Or one of the catchups
-                    if let Some((_, shards_to_download)) =
-                        self.client.catchup_state_syncs.get_mut(&hash)
-                    {
-                        if let Some(part_id) = state_response.part_id() {
-                            self.state_sync_addr.do_send(
-                                StateSyncActorRequests::ReceivedRequestedPart {
-                                    part_id,
-                                    shard_id,
-                                    hash,
-                                },
-                            );
-                        }
-
-                        if let Some(shard_download) = shards_to_download.get_mut(&shard_id) {
-                            assert!(download.is_none(), "Internal downloads set has duplicates");
-                            download = Some(shard_download);
-                        } else {
-                            // This may happen because of sending too many StateRequests to different peers.
-                            // For example, we received StateResponse after StateSync completion.
-                        }
-                    }
-                    // We should not be requesting the same state twice.
-                    download
-                };
-
-                if let Some(shard_sync_download) = download {
-                    match shard_sync_download.status {
-                        ShardSyncStatus::StateDownloadHeader => {
-                            if let Some(header) = state_response.take_header() {
-                                if !shard_sync_download.downloads[0].done {
-                                    match self.client.chain.set_state_header(shard_id, hash, header)
-                                    {
-                                        Ok(()) => {
-                                            shard_sync_download.downloads[0].done = true;
-                                        }
-                                        Err(err) => {
-                                            error!(target: "sync", "State sync set_state_header error, shard = {}, hash = {}: {:?}", shard_id, hash, err);
-                                            shard_sync_download.downloads[0].error = true;
-                                        }
-                                    }
-                                }
-                            } else {
-                                // No header found.
-                                // It may happen because requested node couldn't build state response.
-                                if !shard_sync_download.downloads[0].done {
-                                    info!(target: "sync", "state_response doesn't have header, should be re-requested, shard = {}, hash = {}", shard_id, hash);
-                                    shard_sync_download.downloads[0].error = true;
-                                }
-                            }
-                        }
-                        ShardSyncStatus::StateDownloadParts => {
-                            if let Some(part) = state_response.take_part() {
-                                let num_parts = shard_sync_download.downloads.len() as u64;
-                                let (part_id, data) = part;
-                                if part_id >= num_parts {
-                                    error!(target: "sync", "State sync received incorrect part_id # {:?} for hash {:?}, potential malicious peer", part_id, hash);
-                                    return NetworkClientResponses::NoResponse;
-                                }
-                                if !shard_sync_download.downloads[part_id as usize].done {
-                                    match self
-                                        .client
-                                        .chain
-                                        .set_state_part(shard_id, hash, part_id, num_parts, &data)
-                                    {
-                                        Ok(()) => {
-                                            shard_sync_download.downloads[part_id as usize].done =
-                                                true;
-                                        }
-                                        Err(err) => {
-                                            error!(target: "sync", "State sync set_state_part error, shard = {}, part = {}, hash = {}: {:?}", shard_id, part_id, hash, err);
-                                            shard_sync_download.downloads[part_id as usize].error =
-                                                true;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                } else {
-                    error!(target: "sync", "State sync received hash {} that we're not expecting, potential malicious peer", hash);
-                }
-
+                self.state_sync_addr
+                    .do_send(StateSyncActorRequests::StateResponse(state_response_info));
                 NetworkClientResponses::NoResponse
             }
             NetworkClientMessages::PartialEncodedChunkRequest(part_request_msg, route_back) => {
